# swagger.yaml
openapi: 3.0.0
info:
  title: Nagar Parishad School Leaving Certificate API
  version: 1.0.0
  description: RESTful API for managing Nagar Parishad School Leaving Certificates with audit trail support
  contact:
    name: API Support
    email: support@nagarparishad.gov.in
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.kaamlo.com
    description: Production server
...
paths:
...
  /api/certificates/bulk:
    post:
      summary: Create certificate with school and student data in one request
      tags: [Certificates]
      description: Accepts frontend format (camelCase) and creates school, student, and certificate automatically
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: Certificate data in frontend format (camelCase)
                  properties:
                    phoneNumber:
                      type: string
                    serialNumber:
                      type: string
                    schoolRecognitionNumber:
                      type: string
                    udiseNumber:
                      type: string
                    email:
                      type: string
                    generalRegisterNumber:
                      type: string
                    medium:
                      type: string
                    affiliationNumber:
                      type: string
                    studentId:
                      type: string
                    uidAadhaarNumber:
                      type: string
                    fullName:
                      type: string
                    fatherName:
                      type: string
                    surname:
                      type: string
                    motherName:
                      type: string
                    nationality:
                      type: string
                    motherTongue:
                      type: string
                    religion:
                      type: string
                    caste:
                      type: string
                    subCaste:
                      type: string
                    placeOfBirth:
                      type: string
                    taluka:
                      type: string
                    district:
                      type: string
                    state:
                      type: string
                    country:
                      type: string
                    dateOfBirth:
                      type: string
                      format: date
                    dateOfBirthInWords:
                      type: string
                    previousSchool:
                      type: string
                    previousClass:
                      type: string
                    admissionDate:
                      type: string
                      format: date
                    admissionClass:
                      type: string
                    academicProgress:
                      type: string
                    conduct:
                      type: string
                    leavingDate:
                      type: string
                      format: date
                    leavingClass:
                      type: string
                    classStudyingDetails:
                      type: string
                    reasonForLeaving:
                      type: string
                    remarks:
                      type: string
                    generalRegisterReference:
                      type: string
                    certificateDate:
                      type: string
                      format: date
                    certificateMonth:
                      type: string
                    certificateYear:
                      type: string
                      description: Can be in Marathi numerals (e.g., "२०२४")
                    classTeacherSignature:
                      type: string
                    clerkSignature:
                      type: string
                    headmasterSignature:
                      type: string
                createdBy:
                  type: string
                  description: Username or user ID
      responses:
        '201':
          description: Certificate created successfully with school and student
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/LeavingCertificate'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Certificate with this serial number already exists
...
  /api/audit:
    get:
      summary: Get audit logs with optional filters
      tags: [Audit]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: table_name
          schema:
            type: string
          description: Filter by table name (schools, students, leaving_certificates)
        - in: query
          name: record_id
          schema:
            type: integer
          description: Filter by record ID
        - in: query
          name: changed_by
          schema:
            type: integer
          description: Filter by user ID
        - in: query
          name: action
          schema:
            type: string
            enum: [INSERT, UPDATE, DELETE]
          description: Filter by action type
        - in: query
          name: action_type
          schema:
            type: string
            enum: [view, add, update, delete, login, logout]
          description: Filter by action type (enhanced)
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Items per page
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          description: Start date filter
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          description: End date filter
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Limit number of results
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
  /api/audit/{table_name}/{record_id}:
    get:
      summary: Get audit logs for specific table and record
      tags: [Audit]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table_name
          required: true
          schema:
            type: string
            enum: [schools, students, leaving_certificates]
          description: Table name
        - in: path
          name: record_id
          required: true
          schema:
            type: integer
          description: Record ID
      responses:
        '200':
          description: Audit logs for the record
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
        '401':
          description: Not authenticated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/auth/otp/verify
  schemas:
...
    LeavingCertificate:
      type: object
      required: [school_id, student_id, serial_no, leaving_date, leaving_class]
      properties:
        id:
          type: integer
        school_id:
          type: integer
          description: Reference to schools table
        student_id:
          type: integer
          description: Reference to students table
        serial_no:
          type: string
          example: '101'
        previous_school:
          type: string
        previous_class:
          type: string
        admission_date:
          type: string
          format: date
          example: '2019-06-01'
        admission_class:
          type: string
          example: Class 1
        progress_in_studies:
          type: string
          example: Good
        conduct:
          type: string
          example: Excellent
        leaving_date:
          type: string
          format: date
          example: '2024-03-31'
        leaving_class:
          type: string
          example: Class 5
        studying_class_and_since:
          type: string
          example: Class 5 since June 2023
        reason_for_leaving:
          type: string
          example: Transfer to another school
        remarks:
          type: string
        general_register_ref:
          type: string
          example: GR001
        certificate_date:
          type: string
          format: date
          example: '2024-04-01'
        certificate_month:
          type: string
          example: April
        certificate_year:
          type: integer
          example: 2024
        class_teacher_name:
          type: string
          example: श्रीमती सुनीता पाटिल
        clerk_name:
          type: string
          example: श्री रवींद्र जोशी
        headmaster_name:
          type: string
          example: श्री महेश देशमुख
        status:
          type: string
          enum: [new, in_review, rejected, accepted, draft, issued, archived, cancelled]
          default: new
          description: |
            - new: When created by Admin (can edit)
            - in_review: When submitted for review (can edit)
            - rejected: When rejected by Super Admin (can edit)
            - accepted: When approved by Super Admin (cannot edit, except Super Admin)
            - draft, issued, archived, cancelled: Legacy statuses
        created_by:
          type: integer
        updated_by:
          type: integer
        issued_by:
          type: integer
          description: User ID who issued the certificate
        issued_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        table_name:
          type: string
          example: leaving_certificates
        record_id:
          type: integer
        action:
          type: string
          enum: [INSERT, UPDATE, DELETE]
        action_type:
          type: string
          enum: [view, add, update, delete, login, logout]
          description: Type of action for audit tracking
        location:
          type: string
          description: Geographic location of the action
        field_name:
          type: string
        old_value:
          type: string
        new_value:
          type: string
        changed_by:
          type: integer
        changed_by_username:
          type: string
        changed_by_name:
          type: string
        changed_at:
          type: string
          format: date-time
        ip_address:
          type: string
        user_agent:
          type: string
        notes:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: object
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            errors:
              - msg: Field is required
                param: field_name
...